Refactor Email.js
