+++ src/MainContainer/Email/ColorLibConnector.js
@@ -1,0 +1,26 @@
+import StepConnector from "@material-ui/core/StepConnector";
+import withStyles from "@material-ui/styles/withStyles";
+
+export default withStyles({
+  alternativeLabel: {
+    top: 22
+  },
+  active: {
+    "& $line": {
+      backgroundImage:
+        "linear-gradient( 95deg,rgb(242,113,33) 0%,rgb(233,64,87) 50%,rgb(138,35,135) 100%)"
+    }
+  },
+  completed: {
+    "& $line": {
+      backgroundImage:
+        "linear-gradient( 95deg,rgb(242,113,33) 0%,rgb(233,64,87) 50%,rgb(138,35,135) 100%)"
+    }
+  },
+  line: {
+    height: 3,
+    border: 0,
+    backgroundColor: "#eaeaf0",
+    borderRadius: 1
+  }
+})(StepConnector);
