+++ src/MainContainer/Email/ColorLibStepIcon.js
@@ -1,0 +1,54 @@
+import React from "react";
+import clsx from "clsx";
+import { makeStyles } from "@material-ui/core/styles";
+import GroupAddIcon from "@material-ui/icons/GroupAdd";
+import SentimentVerySatisfiedIcon from "@material-ui/icons/SentimentVerySatisfied";
+import CheckIcon from "@material-ui/icons/Check";
+
+const useColorlibStepIconStyles = makeStyles({
+  root: {
+    backgroundColor: "#ccc",
+    zIndex: 1,
+    color: "#fff",
+    width: 50,
+    height: 50,
+    display: "flex",
+    borderRadius: "50%",
+    justifyContent: "center",
+    alignItems: "center",
+    cursor: "pointer"
+  },
+  active: {
+    backgroundImage:
+      "linear-gradient( 136deg, rgb(242,113,33) 0%, rgb(233,64,87) 50%, rgb(138,35,135) 100%)",
+    boxShadow: "0 4px 10px 0 rgba(0,0,0,.25)"
+  },
+  completed: {
+    backgroundImage:
+      "linear-gradient( 136deg, rgb(242,113,33) 0%, rgb(233,64,87) 50%, rgb(138,35,135) 100%)"
+  }
+});
+
+const ColorlibStepIcon = props => {
+  const classes = useColorlibStepIconStyles();
+  const { active, completed } = props;
+
+  const icons = {
+    1: <CheckIcon />,
+    2: <GroupAddIcon />,
+    3: <SentimentVerySatisfiedIcon />
+  };
+
+  return (
+    <div
+      className={clsx(classes.root, {
+        [classes.active]: active,
+        [classes.completed]: completed
+      })}
+    >
+      {icons[String(props.icon)]}
+    </div>
+  );
+};
+
+export default ColorlibStepIcon;
